cleaned up project structure by moving components to their own folder and pages
